move to git and implement update strategy
There is not a supereasy way to update pkgsrc on Lunix.  `pkg_comp`
uses BSD chroot jails to build updated packages; so it can't be ported
to unprivleged linux.

This is one of the techniques listed in how to upgrade packages
http://wiki.netbsd.org/pkgsrc/how_to_upgrade_packages/

I'm using the git checksum of the checked out pkgsrc directory in the
prefix that I use on the inital bootstrap command; and then I create
a symbolic link to it from $HOME/pkg -> $HOME/pkgbld/<gitchecksum>

To update; use git to update the pkgsrc tree; then create a new
$HOME/pkgbld/<gitchecksum>.  Once everthing is rebuilt in the new place;
then swap out the symlink.  Run your application test suite.  If your
tests don't pass; roll back by swapping the symlink back to the last
good build.